Add the Edit menu.
It is only active when the EditDialog is up, so I also added some
methods to EditDialog to connect and to disconnect the menu actions
to the dialog's QPlainTextEdit. It turns out that copy and paste
works in that widget without the menu connections, but this makes
the connection to main window's Edit menu explicit.--The menu won't
work without the connections, for instance.

Since the actions only make sense in the context of the EditDialog,
the code disabled the actions in the MainWindow.

Signed-off-by: Jason Stephenson <jason@sigio.com>