Add ability to save as a fortune file
The program can now save as a delimited string file for the fortune
program. It uses the standard delimiter of '%' found on a line by
itself.

The file name is used to determine if the data should be written as an
Omifile or as a Strfile.  If the filename ends with ".omi," then an
Omifile is written, otherwise a Strfile is output.

Signed-off-by: Jason Stephenson <jason@sigio.com>