Fix signals and slots issues by simplifying things
Drop the signals from OmiDoc and replace them with just two that are
emitted after a file is loaded to pass the model data to the
MainWindow view.

Simplify a number of the slots in MainWindow and move them to private
funtions.  They still emit the same signals.

Also fix a bug related to setting the MainWindow's filename when a
file is opened.

The controller should probably take over more of the view's
responsibilities.

Signed-off-by: Jason Stephenson <jason@sigio.com>