Wire up some more signals and slots for the FindDialog
Connect the state changed signals of the 5 checkboxes to a new slot
that, in turn, emits one of 5 new signals indicating which checkbox
has changed state and what that state is.

The signal for the "Search From Start" checkbox is wired up to a new
slot on the main window to toggle the state of the isNewSearch
variable.  If the user checks the checkbox in the middle of a search,
we want to treat it as a new search.

This gets us closer to a generic FindDialog implementation that could
go into its own library.

Signed-off-by: Jason Stephenson <jason@sigio.com>