Rewire FindDialog::addSearchTextItem() as a public slot
Change the FindDialog::addSearchTextItem function from private to a
public slot and do not call it every time the find button is clicked
in the dialog.

Add the MainWindow::searchTextFound signal and wire it up to the
FindDialog::addSearchTextItem slot.  Now, we only add successful
search terms to the QComboBox items list.

Signed-off-by: Jason Stephenson <jason@sigio.com>