Add signal and slot to FindDialog for ComboBox changes
Add the FindDialog::comboBoxCurrentTextChanged slot and wire it up to
the QComboBox::currentTextChanged signal.  The new slot emits the
FindDialog::searchTextComboBoxTextChanged signal as way to signal that
the search combobox text has changed.

We also wire this new signal up to a lambda in the main window to set
the isNewSearch variable to true.

Signed-off-by: Jason Stephenson <jason@sigio.com>