Improve the "intelligence" of adding comments and fortunes
If the comment or fortune list has a current item when the add button
is clicked, then the new comment or future is inserted after the
current selection.  If there is no current selection, then the new
comment or fortune is appended to the end of the list.

New signals were added to the main window to indicate that a comment
or fortune was inserted with a given index.  These signals are wired
up to new slots in the OmiDoc controller to put the new in the proper
place in the model.

Signed-off-by: Jason Stephenson <jason@sigio.com>