Actually fix the % bug when reading strfiles
Search the input for "\n%\n" when reading from a fortune Strfile to
avoid splitting on % at the end of a line when it is not on a line by
itself.  This required some adjustment to the length and start
position calculations.

My previous attempt to fix this issue caused and infinite loop over
the input data.

Signed-off-by: Jason Stephenson <jason@sigio.com>