Add comment about beep not working on Linux
QApplication::beep() doesn't do anything in GUI applications on
GNU/Linux, so I added a TODO comment to fix it.

I did some research into "flashing" the text combo box background, and
it seems that I need to do something with the palette to get the
current color, change it, then change it back. I may want to do a
calculation based on the current color to darken or lighten it in case
the user has any special themes going on. (Does this application even
support themes? That's something else to look into.)

Signed-off-by: Jason Stephenson <jason@sigio.com>