Fix #1 .omi files saved when built with Qt 6 cannot be read
This was embarrassingly simple.

The endian conversion functions in Qt are template functions. The
conversions to big endian were not working properly with Qt 6.5
without the template parameter to specify the quint32 type for the
conversion. Adding the template parameter resolves the issue saving
the file.

To be totally correct, this commit also adds the parameter to the uses
of qFromBigEndian.

For some reason it worked without the template parameter on Qt 5.15.

Signed-off-by: Jason Stephenson <jason@sigio.com>